Rename ReactNativeSetNativeProps.js to be more general
